Add infrastructure for changing acquisition function optimizers
The BotorchStrategy is expanded to modularize the acquisition function
optimizer. Every acquisition function optimizer has to derive from a
newly created base class. The current Botorch-based optimizer is moved
into a separate class.

This includes a breaking change in the data model: A new data model for
acquisition function optimizers is created. The Botorch-optimizer params
stored in the BotorchStrategy are moved to the new BotorchAcqfOptimizer.